Create Grace_week_2.md
---
title: "Grace's post week 2"
date: 2021-09-02T00:00:00-04:00
draft: false
---

This is a sample post. You can copy the text of this post to create new posts for your weekly personal and group reports. Pay close attention to the format of the text in this file—changing it too drastically will cause the post to not show up.