Ethan's Week 2 Post
During this week I learned how to employ new technologies relating to content management. Through the use of GitHub I have seen how and why this site is so often used for content management for groups. Through the process of creating a GitHub account and learning about how to make posts such as this one I feel that I have definitely made a step in learning about how to manage content efficiently in a group and learning how to employ technologies related to content management.